reorg route directory structure
