+++ ExpressAPI/components/automobiles.ts
@@ -1,10 +1,9 @@
 import express from 'express'
 
-const app = express()
 const router = express.Router()
 
 router.get('/', (req: any, res: any) => {
-  res.send('uuuSSSeeeRRRsss...uuuSSSeeeRRRsss...uuuSSSeeeRRRsss')
+  res.send('CARS...cars...CARS')
 })
 
 export default router
