+++ ExpressAPI/components/contracts.ts
@@ -1,10 +1,9 @@
 import express from 'express'
 
-const app = express()
 const router = express.Router()
 
 router.get('/', (req: any, res: any) => {
-  res.send('CREDITcardSYSTEMS...CREDITcardSYSTEMS...CREDITcardSYSTEMS')
+  res.send('cOnTrAcTs...cOnTrAcTs...cOnTrAcTs')
 })
 
 export default router
