+++ ExpressAPI/components/creditCardSystems.ts
@@ -1,10 +1,9 @@
 import express from 'express'
 
-const app = express()
 const router = express.Router()
 
 router.get('/', (req: any, res: any) => {
-  res.send('CARS...cars...CARS')
+  res.send('CREDITcardSYSTEMS...CREDITcardSYSTEMS...CREDITcardSYSTEMS')
 })
 
 export default router
