+++ ExpressAPI/components/users.ts
@@ -1,10 +1,9 @@
 import express from 'express'
 
-const app = express()
 const router = express.Router()
 
 router.get('/', (req: any, res: any) => {
-  res.send('cOnTrAcTs...cOnTrAcTs...cOnTrAcTs')
+  res.send('uuuSSSeeeRRRsss...uuuSSSeeeRRRsss...uuuSSSeeeRRRsss')
 })
 
 export default router
