+++ ExpressAPI/server.ts
@@ -1,11 +1,7 @@
 import express from 'express'
 import dotenv from 'dotenv'
 
-import home from './home'
-import automobiles from './automobiles'
-import contracts from './contracts'
-import creditCardSystems from './creditCardSystems'
-import users from './users'
+import router from './components/router'
 
 dotenv.config()
 
@@ -14,11 +10,7 @@ const port = process.env.PORT
 
 const app = express()
 
-app.use('/', home)
-app.use('/automobiles', automobiles)
-app.use('/contracts', contracts)
-app.use('/creditCardSystems', creditCardSystems)
-app.use('/users', users)
+app.use(router)
 
 app.listen(port, () => {
   console.log(`oPay API listening on ${host}${port}`)
